pip-requirements.txt
Beaker==1.3.1
FormEncode==1.2.2
Mako==0.2.4
MySQL-python==1.2.3c1
Paste==1.7.2
PasteDeploy==1.3.3
PasteScript==1.7.3
PyYAML==3.08
Pygments==1.0
Pylons==0.9.7
Routes==1.10.3
SQLAlchemy==0.5.4p2
Shapely==1.0.12
Tempita==0.4
WebError==0.10.1
WebHelpers==0.6.4
WebOb==0.9.6.1
WebTest==1.2
decorator==3.0.1
elementtree==1.2.7-20070827-preview
-e svn+http://geopy.googlecode.com/svn/trunk@75#egg=geopy-0.93dev_r75-py2.5-dev
lxml==2.2.2
meld3==0.6.5
nose==0.11.1
-e hg+https://bitbucket.org/slinkp/whoismygov/@tip#egg=whoismygov-tip
psycopg2==2.0.11
pyproj==1.8.5
python-votesmart==0.2.1
simplejson==2.0.9
sqlalchemy-migrate==0.5.4
supervisor==3.0a7
wsgiref==0.1.2
